+++ x/auth/migrations/v2/store_test.go
@@ -73,10 +73,9 @@ func TestMigrateVestingAccounts(t *testing.T) {
 	legacySubspace := newMockSubspace(authtypes.DefaultParams())
 	require.NoError(t, v4.Migrate(ctx, storeService, legacySubspace, cdc))
 
-	ctx = app.BaseApp.NewContext(false, cmtproto.Header{Time: time.Now()})
+	ctx = app.BaseApp.NewContextLegacy(false, cmtproto.Header{Time: time.Now()})
 	err = stakingKeeper.SetParams(ctx, stakingtypes.DefaultParams())
 	require.NoError(t, err)
-
 	lastAccNum := uint64(1000)
 	createBaseAccount := func(addr sdk.AccAddress) *authtypes.BaseAccount {
 		baseAccount := authtypes.NewBaseAccountWithAddress(addr)