+++ x/auth/tx/legacy_amino_json_test.go
@@ -8,6 +8,8 @@ import (
 	"cosmossdk.io/x/tx/signing/aminojson"
 	"github.com/stretchr/testify/require"
 
+	"github.com/cosmos/cosmos-sdk/codec"
+	"github.com/cosmos/cosmos-sdk/codec/legacy"
 	cdctypes "github.com/cosmos/cosmos-sdk/codec/types"
 	"github.com/cosmos/cosmos-sdk/testutil/testdata"
 	sdk "github.com/cosmos/cosmos-sdk/types"
@@ -38,6 +40,7 @@ func buildTx(t *testing.T, bldr *wrapper) {
 }
 
 func TestLegacyAminoJSONHandler_GetSignBytes(t *testing.T) {
+	legacytx.RegressionTestingAminoCodec = codec.NewLegacyAmino()
 	var (
 		chainID        = "test-chain"
 		accNum  uint64 = 7
@@ -152,6 +155,9 @@ func TestLegacyAminoJSONHandler_AllGetSignBytesComparison(t *testing.T) {
 
 	modeHandler := aminojson.NewSignModeHandler(aminojson.SignModeHandlerOptions{})
 	mode, _ := signing.APISignModeToInternal(modeHandler.Mode())
+	legacyAmino := codec.NewLegacyAmino()
+	legacy.RegisterAminoMsg(legacyAmino, &banktypes.MsgSend{}, "cosmos-sdk/MsgSend")
+	legacytx.RegressionTestingAminoCodec = legacyAmino
 
 	testcases := []struct {
 		name           string
@@ -221,6 +227,8 @@ func TestLegacyAminoJSONHandler_AllGetSignBytesComparison(t *testing.T) {
 			require.Equal(t, string(tc.expectedSignBz), string(newSignBz))
 		})
 	}
+
+	legacytx.RegressionTestingAminoCodec = nil
 }
 
 func TestLegacyAminoJSONHandler_DefaultMode(t *testing.T) {