+++ x/evidence/genesis_test.go
@@ -5,6 +5,7 @@ import (
 	"testing"
 	"time"
 
+	"cosmossdk.io/collections"
 	"cosmossdk.io/depinject"
 	"cosmossdk.io/log"
 	"cosmossdk.io/x/evidence"
@@ -39,7 +40,7 @@ func (suite *GenesisTestSuite) SetupTest() {
 		&evidenceKeeper)
 	require.NoError(suite.T(), err)
 
-	suite.ctx = app.BaseApp.NewContext(false, cmtproto.Header{Height: 1})
+	suite.ctx = app.BaseApp.NewContextLegacy(false, cmtproto.Header{Height: 1})
 	suite.keeper = evidenceKeeper
 }
 
@@ -73,7 +74,7 @@ func (suite *GenesisTestSuite) TestInitGenesis() {
 			true,
 			func() {
 				for _, e := range testEvidence {
-					_, err := suite.keeper.GetEvidence(suite.ctx, e.Hash())
+					_, err := suite.keeper.Evidences.Get(suite.ctx, e.Hash())
 					suite.NoError(err)
 				}
 			},
@@ -93,9 +94,8 @@ func (suite *GenesisTestSuite) TestInitGenesis() {
 			},
 			false,
 			func() {
-				ev, err := suite.keeper.GetAllEvidence(suite.ctx)
-				suite.Empty(ev)
-				suite.NoError(err)
+				_, err := suite.keeper.Evidences.Iterate(suite.ctx, nil)
+				suite.Require().ErrorIs(err, collections.ErrInvalidIterator)
 			},
 		},
 	}
@@ -133,13 +133,14 @@ func (suite *GenesisTestSuite) TestExportGenesis() {
 		{
 			"success",
 			func() {
-				err := suite.keeper.SetEvidence(suite.ctx, &types.Equivocation{
+				ev := &types.Equivocation{
 					Height:           1,
 					Power:            100,
 					Time:             time.Now().UTC(),
 					ConsensusAddress: pk.PubKey().Address().String(),
-				})
-				suite.NoError(err)
+				}
+				suite.Require().NoError(suite.keeper.Evidences.Set(suite.ctx, ev.Hash(), ev))
+
 			},
 			true,
 			func() {},